Enable clang-format, restart VkToyApp from scratch to follow tutorial
